Extraer modal de "Mis marcaciones" de mark.js
Cuarto paso de la descomposición de mark.js: openMyEventsModal() y
closeMyEventsModal() se mueven a mark/my-events-modal.js. Es el modal
más autocontenido del archivo — no comparte estado mutable con el
resto de mark.js (a diferencia del modal de éxito/error, que sí
depende de errorModalVisible/previousActiveElement compartidos con el
resto del flujo de identificación). Refactor puro, sin cambios de
comportamiento.
